rewrote such that certain parts can be reused when changing vehicle settings
