+++ pages/ev/index.tsx
@@ -1,10 +1,10 @@
-import EVMakeList from '@/components/ev/EVMakeList'
+import AddVehicle from '@/components/ev/AddVehicle'
 import Layout from '@/src/layout'
 
 export default function EV() {
     return (
         <Layout>
-            <EVMakeList />
+            <AddVehicle />
         </Layout>
     )
 }
