+++ src/components/ev/charger/charger.module.css
@@ -1,0 +1,11 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
+
+.radio__container {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
