+++ src/components/ev/charger/texts.ts
@@ -1,0 +1,14 @@
+export const texts = {
+    first: {
+        nb: 'Hva er din',
+        en: 'What is your',
+    },
+    second: {
+        nb: 'ladekapasitet?',
+        en: 'charger capacity?',
+    },
+    info: {
+        nb: 'Kapasiteten til din hjemmelader. De fleste har 3.7 kW.',
+        en: 'Capacity of your charger at home. Most people have 3.7 kW.',
+    },
+}
