+++ src/components/ev/interval/interval.module.css
@@ -1,0 +1,5 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
