+++ src/components/ev/interval/texts.ts
@@ -1,0 +1,14 @@
+export const texts = {
+    first: {
+        nb: 'Hva er ditt',
+        en: 'What is your',
+    },
+    second: {
+        nb: 'lademønster?',
+        en: 'charging pattern?',
+    },
+    info: {
+        nb: 'Batterinivået ved start og stopp vil hjelpe oss å regne ut et presist ladevindu.',
+        en: 'Battery level when you start charging and the target level will helps us calculate the length of the charging window.',
+    },
+}
