+++ src/components/ev/model/texts.ts
@@ -1,0 +1,14 @@
+export const texts = {
+    first: {
+        nb: 'Hvilken',
+        en: 'Which',
+    },
+    second: {
+        nb: 'bil',
+        en: 'car',
+    },
+    third: {
+        nb: 'har du?',
+        en: 'do you have?',
+    },
+}
