+++ src/components/ev/style.module.css
@@ -17,11 +17,6 @@
     gap: 1.5rem;
 }
 
-.radio__container {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-}
 
 .onboarding__controls {
     width: 100%;
