+++ src/components/ev/texts.ts
@@ -17,48 +17,4 @@ export const texts = {
             en: 'Generate charging plan',
         },
     },
-    headers: {
-        make: {
-            first: {
-                nb: 'Hvilken',
-                en: 'Which',
-            },
-            second: {
-                nb: 'bil',
-                en: 'car',
-            },
-            third: {
-                nb: 'har du?',
-                en: 'do you have?',
-            },
-        },
-        charger: {
-            first: {
-                nb: 'Hva er din',
-                en: 'What is your',
-            },
-            second: {
-                nb: 'ladekapasitet?',
-                en: 'charger capacity?',
-            },
-            info: {
-                nb: 'Kapasiteten til din hjemmelader. De fleste har 3.7 kW.',
-                en: 'Capacity of your charger at home. Most people have 3.7 kW.',
-            },
-        },
-        interval: {
-            first: {
-                nb: 'Hva er ditt',
-                en: 'What is your',
-            },
-            second: {
-                nb: 'lademønster?',
-                en: 'charging pattern?',
-            },
-            info: {
-                nb: 'Batterinivået ved start og stopp vil hjelpe oss å regne ut et presist ladevindu.',
-                en: 'Battery level when you start charging and the target level will helps us calculate the length of the charging window.',
-            },
-        },
-    },
 }
